Move code action UI into controller
This merges the codeActionUi class into `CodeActionController` to reduce the number of abstractions
