+++ src/vs/editor/contrib/codeAction/browser/codeActionController.ts
@@ -3,24 +3,39 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
+import { getDomNodePagePosition } from 'vs/base/browser/dom';
 import { IAnchor } from 'vs/base/browser/ui/contextview/contextview';
+import { IAction } from 'vs/base/common/actions';
+import { onUnexpectedError } from 'vs/base/common/errors';
 import { Lazy } from 'vs/base/common/lazy';
-import { Disposable } from 'vs/base/common/lifecycle';
+import { Disposable, MutableDisposable } from 'vs/base/common/lifecycle';
 import { ICodeEditor } from 'vs/editor/browser/editorBrowser';
-import { IPosition } from 'vs/editor/common/core/position';
-import { IEditorContribution } from 'vs/editor/common/editorCommon';
+import { IPosition, Position } from 'vs/editor/common/core/position';
+import { IEditorContribution, ScrollType } from 'vs/editor/common/editorCommon';
 import { CodeActionTriggerType } from 'vs/editor/common/languages';
 import { ILanguageFeaturesService } from 'vs/editor/common/services/languageFeatures';
-import { applyCodeAction, ApplyCodeActionReason } from 'vs/editor/contrib/codeAction/browser/codeAction';
-import { CodeActionUi } from 'vs/editor/contrib/codeAction/browser/codeActionUi';
+import { ApplyCodeActionReason, applyCodeAction } from 'vs/editor/contrib/codeAction/browser/codeAction';
+import { CodeActionKeybindingResolver } from 'vs/editor/contrib/codeAction/browser/codeActionKeybindingResolver';
+import { toMenuItems } from 'vs/editor/contrib/codeAction/browser/codeActionMenu';
+import { LightBulbWidget } from 'vs/editor/contrib/codeAction/browser/lightBulbWidget';
 import { MessageController } from 'vs/editor/contrib/message/browser/messageController';
+import { localize } from 'vs/nls';
+import { IActionListDelegate } from 'vs/platform/actionWidget/browser/actionList';
+import { IActionWidgetService } from 'vs/platform/actionWidget/browser/actionWidget';
+import { ICommandService } from 'vs/platform/commands/common/commands';
+import { IConfigurationService } from 'vs/platform/configuration/common/configuration';
 import { IContextKeyService } from 'vs/platform/contextkey/common/contextkey';
 import { IInstantiationService } from 'vs/platform/instantiation/common/instantiation';
 import { IMarkerService } from 'vs/platform/markers/common/markers';
 import { IEditorProgressService } from 'vs/platform/progress/common/progress';
 import { CodeActionAutoApply, CodeActionFilter, CodeActionItem, CodeActionSet, CodeActionTrigger, CodeActionTriggerSource } from '../common/types';
-import { CodeActionModel } from './codeActionModel';
+import { CodeActionModel, CodeActionsState } from './codeActionModel';
 
+
+interface IActionShowOptions {
+	readonly includeDisabledActions?: boolean;
+	readonly fromLightbulb?: boolean;
+}
 
 export class CodeActionController extends Disposable implements IEditorContribution {
 
@@ -32,40 +47,54 @@ export class CodeActionController extends Disposable implements IEditorContribut
 
 	private readonly _editor: ICodeEditor;
 	private readonly _model: CodeActionModel;
-	private readonly _ui: Lazy<CodeActionUi>;
+
+	private readonly _lightBulbWidget: Lazy<LightBulbWidget | null>;
+	private readonly _activeCodeActions = this._register(new MutableDisposable<CodeActionSet>());
+	private _showDisabled = false;
+
+	private readonly _resolver: CodeActionKeybindingResolver;
+
+	#disposed = false;
 
 	constructor(
 		editor: ICodeEditor,
 		@IMarkerService markerService: IMarkerService,
 		@IContextKeyService contextKeyService: IContextKeyService,
+		@IInstantiationService instantiationService: IInstantiationService,
+		@ILanguageFeaturesService languageFeaturesService: ILanguageFeaturesService,
 		@IEditorProgressService progressService: IEditorProgressService,
+		@ICommandService private readonly _commandService: ICommandService,
+		@IConfigurationService private readonly _configurationService: IConfigurationService,
+		@IActionWidgetService private readonly _actionWidgetService: IActionWidgetService,
 		@IInstantiationService private readonly _instantiationService: IInstantiationService,
-		@ILanguageFeaturesService languageFeaturesService: ILanguageFeaturesService
 	) {
 		super();
 
 		this._editor = editor;
 
 		this._model = this._register(new CodeActionModel(this._editor, languageFeaturesService.codeActionProvider, markerService, contextKeyService, progressService));
+		this._register(this._model.onDidChangeState(newState => this.update(newState)));
 
-		this._register(this._model.onDidChangeState(newState => this._ui.value.update(newState)));
-
-		this._ui = new Lazy(() => this._register(_instantiationService.createInstance(CodeActionUi, editor, {
-			applyCodeAction: async (action, retrigger, preview) => {
-				try {
-					await this._applyCodeAction(action, preview);
-				} finally {
-					if (retrigger) {
-						this._trigger({ type: CodeActionTriggerType.Auto, triggerAction: CodeActionTriggerSource.QuickFix, filter: {} });
-					}
-				}
+		this._lightBulbWidget = new Lazy(() => {
+			const widget = this._editor.getContribution<LightBulbWidget>(LightBulbWidget.ID);
+			if (widget) {
+				this._register(widget.onClick(e => this.showCodeActionList(e.actions, e, { includeDisabledActions: false, fromLightbulb: true })));
 			}
-		}))
-		);
+			return widget;
+		});
+
+		this._resolver = instantiationService.createInstance(CodeActionKeybindingResolver);
+
+		this._register(this._editor.onDidLayoutChange(() => this._actionWidgetService.hide()));
+	}
+
+	override dispose() {
+		this.#disposed = true;
+		super.dispose();
 	}
 
 	public showCodeActions(_trigger: CodeActionTrigger, actions: CodeActionSet, at: IAnchor | IPosition) {
-		return this._ui.value.showCodeActionList(actions, at, { includeDisabledActions: false, fromLightbulb: false });
+		return this.showCodeActionList(actions, at, { includeDisabledActions: false, fromLightbulb: false });
 	}
 
 	public manualTriggerAtCurrentPosition(
@@ -87,7 +116,206 @@ export class CodeActionController extends Disposable implements IEditorContribut
 		return this._model.trigger(trigger);
 	}
 
-	private _applyCodeAction(action: CodeActionItem, preview: boolean): Promise<void> {
-		return this._instantiationService.invokeFunction(applyCodeAction, action, ApplyCodeActionReason.FromCodeActions, { preview, editor: this._editor });
+	private async _applyCodeAction(action: CodeActionItem, retrigger: boolean, preview: boolean): Promise<void> {
+		try {
+			await this._instantiationService.invokeFunction(applyCodeAction, action, ApplyCodeActionReason.FromCodeActions, { preview, editor: this._editor });
+		} finally {
+			if (retrigger) {
+				this._trigger({ type: CodeActionTriggerType.Auto, triggerAction: CodeActionTriggerSource.QuickFix, filter: {} });
+			}
+		}
+	}
+
+	private async update(newState: CodeActionsState.State): Promise<void> {
+		if (newState.type !== CodeActionsState.Type.Triggered) {
+			this._lightBulbWidget.rawValue?.hide();
+			return;
+		}
+
+		let actions: CodeActionSet;
+		try {
+			actions = await newState.actions;
+		} catch (e) {
+			onUnexpectedError(e);
+			return;
+		}
+
+		if (this.#disposed) {
+			return;
+		}
+
+		this._lightBulbWidget.value?.update(actions, newState.trigger, newState.position);
+
+		if (newState.trigger.type === CodeActionTriggerType.Invoke) {
+			if (newState.trigger.filter?.include) { // Triggered for specific scope
+				// Check to see if we want to auto apply.
+
+				const validActionToApply = this.tryGetValidActionToApply(newState.trigger, actions);
+				if (validActionToApply) {
+					try {
+						this._lightBulbWidget.value?.hide();
+						await this._applyCodeAction(validActionToApply, false, false);
+					} finally {
+						actions.dispose();
+					}
+					return;
+				}
+
+				// Check to see if there is an action that we would have applied were it not invalid
+				if (newState.trigger.context) {
+					const invalidAction = this.getInvalidActionThatWouldHaveBeenApplied(newState.trigger, actions);
+					if (invalidAction && invalidAction.action.disabled) {
+						MessageController.get(this._editor)?.showMessage(invalidAction.action.disabled, newState.trigger.context.position);
+						actions.dispose();
+						return;
+					}
+				}
+			}
+
+			const includeDisabledActions = !!newState.trigger.filter?.include;
+			if (newState.trigger.context) {
+				if (!actions.allActions.length || !includeDisabledActions && !actions.validActions.length) {
+					MessageController.get(this._editor)?.showMessage(newState.trigger.context.notAvailableMessage, newState.trigger.context.position);
+					this._activeCodeActions.value = actions;
+					actions.dispose();
+					return;
+				}
+			}
+
+			this._activeCodeActions.value = actions;
+			this.showCodeActionList(actions, this.toCoords(newState.position), { includeDisabledActions, fromLightbulb: false });
+		} else {
+			// auto magically triggered
+			if (this._actionWidgetService.isVisible) {
+				// TODO: Figure out if we should update the showing menu?
+				actions.dispose();
+			} else {
+				this._activeCodeActions.value = actions;
+			}
+		}
+	}
+
+	private getInvalidActionThatWouldHaveBeenApplied(trigger: CodeActionTrigger, actions: CodeActionSet): CodeActionItem | undefined {
+		if (!actions.allActions.length) {
+			return undefined;
+		}
+
+		if ((trigger.autoApply === CodeActionAutoApply.First && actions.validActions.length === 0)
+			|| (trigger.autoApply === CodeActionAutoApply.IfSingle && actions.allActions.length === 1)
+		) {
+			return actions.allActions.find(({ action }) => action.disabled);
+		}
+
+		return undefined;
+	}
+
+	private tryGetValidActionToApply(trigger: CodeActionTrigger, actions: CodeActionSet): CodeActionItem | undefined {
+		if (!actions.validActions.length) {
+			return undefined;
+		}
+
+		if ((trigger.autoApply === CodeActionAutoApply.First && actions.validActions.length > 0)
+			|| (trigger.autoApply === CodeActionAutoApply.IfSingle && actions.validActions.length === 1)
+		) {
+			return actions.validActions[0];
+		}
+
+		return undefined;
+	}
+
+	public async showCodeActionList(actions: CodeActionSet, at: IAnchor | IPosition, options: IActionShowOptions): Promise<void> {
+		const editorDom = this._editor.getDomNode();
+		if (!editorDom) {
+			return;
+		}
+
+		const actionsToShow = options.includeDisabledActions && (this._showDisabled || actions.validActions.length === 0) ? actions.allActions : actions.validActions;
+		if (!actionsToShow.length) {
+			return;
+		}
+
+		const anchor = Position.isIPosition(at) ? this.toCoords(at) : at;
+
+		const delegate: IActionListDelegate<CodeActionItem> = {
+			onSelect: async (action: CodeActionItem, preview?: boolean) => {
+				this._applyCodeAction(action, /* retrigger */ true, !!preview);
+				this._actionWidgetService.hide();
+			},
+			onHide: () => {
+				this._editor?.focus();
+			}
+		};
+
+		this._actionWidgetService.show(
+			'codeActionWidget',
+			true,
+			toMenuItems(actionsToShow, this._shouldShowHeaders(), this._resolver.getResolver()),
+			delegate,
+			anchor,
+			editorDom,
+			this._getActionBarActions(actions, at, options));
+	}
+
+	private toCoords(position: IPosition): IAnchor {
+		if (!this._editor.hasModel()) {
+			return { x: 0, y: 0 };
+		}
+
+		this._editor.revealPosition(position, ScrollType.Immediate);
+		this._editor.render();
+
+		// Translate to absolute editor position
+		const cursorCoords = this._editor.getScrolledVisiblePosition(position);
+		const editorCoords = getDomNodePagePosition(this._editor.getDomNode());
+		const x = editorCoords.left + cursorCoords.left;
+		const y = editorCoords.top + cursorCoords.top + cursorCoords.height;
+
+		return { x, y };
+	}
+
+	private _shouldShowHeaders(): boolean {
+		const model = this._editor?.getModel();
+		return this._configurationService.getValue('editor.codeActionWidget.showHeaders', { resource: model?.uri });
+	}
+
+	private _getActionBarActions(actions: CodeActionSet, at: IAnchor | IPosition, options: IActionShowOptions): IAction[] {
+		if (options.fromLightbulb) {
+			return [];
+		}
+
+		const resultActions = actions.documentation.map((command): IAction => ({
+			id: command.id,
+			label: command.title,
+			tooltip: command.tooltip ?? '',
+			class: undefined,
+			enabled: true,
+			run: () => this._commandService.executeCommand(command.id, ...(command.arguments ?? [])),
+		}));
+
+		if (options.includeDisabledActions && actions.validActions.length > 0 && actions.allActions.length !== actions.validActions.length) {
+			resultActions.push(this._showDisabled ? {
+				id: 'hideMoreActions',
+				label: localize('hideMoreActions', 'Hide Disabled'),
+				enabled: true,
+				tooltip: '',
+				class: undefined,
+				run: () => {
+					this._showDisabled = false;
+					return this.showCodeActionList(actions, at, options);
+				}
+			} : {
+				id: 'showMoreActions',
+				label: localize('showMoreActions', 'Show Disabled'),
+				enabled: true,
+				tooltip: '',
+				class: undefined,
+				run: () => {
+					this._showDisabled = true;
+					return this.showCodeActionList(actions, at, options);
+				}
+			});
+		}
+
+		return resultActions;
 	}
 }
