+++ src/vs/editor/contrib/codeAction/browser/lightBulbWidget.ts
@@ -14,7 +14,6 @@ import 'vs/css!./lightBulbWidget';
 import { ContentWidgetPositionPreference, ICodeEditor, IContentWidget, IContentWidgetPosition } from 'vs/editor/browser/editorBrowser';
 import { EditorOption } from 'vs/editor/common/config/editorOptions';
 import { IPosition } from 'vs/editor/common/core/position';
-import { IEditorContribution } from 'vs/editor/common/editorCommon';
 import { computeIndentLevel } from 'vs/editor/common/model/utils';
 import { autoFixCommandId, quickFixCommandId } from 'vs/editor/contrib/codeAction/browser/codeAction';
 import type { CodeActionSet, CodeActionTrigger } from 'vs/editor/contrib/codeAction/common/types';
@@ -44,7 +43,7 @@ namespace LightBulbState {
 	export type State = typeof Hidden | Showing;
 }
 
-export class LightBulbWidget extends Disposable implements IEditorContribution, IContentWidget {
+export class LightBulbWidget extends Disposable implements IContentWidget {
 
 	public static readonly ID = 'editor.contrib.lightbulbWidget';
 
@@ -52,7 +51,7 @@ export class LightBulbWidget extends Disposable implements IEditorContribution, 
 
 	private readonly _domNode: HTMLElement;
 
-	private readonly _onClick = this._register(new Emitter<{ x: number; y: number; actions: CodeActionSet; trigger: CodeActionTrigger }>());
+	private readonly _onClick = this._register(new Emitter<{ readonly x: number; readonly y: number; readonly actions: CodeActionSet; readonly trigger: CodeActionTrigger }>());
 	public readonly onClick = this._onClick.event;
 
 	private _state: LightBulbState.State = LightBulbState.Hidden;
